% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_coalesce.R
\name{data_coalesce}
\alias{data_coalesce}
\title{Title}
\usage{
data_coalesce(ledger, input_custom)
}
\arguments{
\item{ledger}{ledger}

\item{input_custom}{vector of columns to be coalesced with their custom counterpart}
}
\value{
ledger
}
\description{
Title
}
