% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_plot.R
\name{hist_plot}
\alias{hist_plot}
\title{Title}
\usage{
hist_plot(short_ledger, time_key, savings)
}
\arguments{
\item{short_ledger}{sl df}

\item{time_key}{see}

\item{future_net}{future net amount per month}
}
\description{
Title
}
