% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_plot.R
\name{net_plot}
\alias{net_plot}
\title{Title}
\usage{
net_plot(ledger, time_key)
}
\arguments{
\item{ledger}{short ledger as df}

\item{time_key}{date key in short ledger: year, quarter, month, date}
}
\description{
Title
}
